Refactor, remove eslint errors.
