+++ app/react-app/src/components/CustomerInfoForm/index.js
@@ -2,8 +2,8 @@ import React, { PropTypes, Component } from 'react';
 import { Link } from 'react-router';
 import { Field, reduxForm } from 'redux-form';
 import { FlatButton } from 'material-ui';
-import Input from './Input';
-import './CustomerInfoForm.css';
+import Input from '../../components/Input';
+import './styles.css';
 
 class CustomerInfoForm extends Component {
