+++ app/react-app/src/components/Header/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './Header.css'
+import './styles.css'
 
 const Header = () => {
     const title = 'Welcome to the atsea shop'
