+++ app/react-app/src/components/Product/index.js
@@ -1,8 +1,7 @@
 import React, { Component, PropTypes } from 'react'
-import './Product.css'
+import './styles.css'
 
 class Product extends Component {
-        // {name} - &#36;{price}{quantity ? ` x ${quantity}` : null}
   render() {
     const { price, quantity, name, image } = this.props;
     const image2 = (
